Fix loading OpenSSL on macOS
On MacOS, the OpenSSL version required by Golem needs to be installed
via homebrew. Golem relies on a fixed path to where the library is
installed. As part of upgrade to OpenSSL version 1.1 the installation
path has changed. On newly set up systems this prevented Golem from
finding the library. This commit appends the new installation path
to the lookup array.